+++ changelog.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [v0.4] - 2019-05-12
+
+- Adding new methods and functions
+- Improving ```Keyboard Driver```
+    - <small>New ```Numeral Keys```</small>
+    - <small>New ```Function Keys``` from ```F1``` to ```F5```</small>
+- Implementing a new ```Applications``` namespace
+- Implementing a Singleton Pattern for Shiro (kernel)
+    - <small>The ```(auto &shiro)``` parameter is no longer required when calling a method</small>
+    - <small>Now you can call ```System::Shiro::GetInstance()``` to access Shiro's methods</small>
+- Implementing a Singleton Pattern for Shell (application)
+    - <small>Now you can call ```Applications::Shell::GetInstance()``` to access the Shell's methods</small>
+- Implementing a ```WatterMark Allocator``` for ```Memory Management```
+    - <small>This is a very (very, very, ..., very) basic version of the feature</small>
+- Implementing the ```new``` operator
+- Implementing a Shell Switcher
+    - <small>You can now change terminals by pressing ```F1``` to ```F5```</small>
+    - <small>```F5``` is the ```Debug Terminal```</small>
+    - <small>The current terminal is displayed at the top of the screen (header)</small>
+- Removing ```VGA.Start()``` from inside ```Modes::Text.Start()```
+    - <small>Now you can find it at ```Shiro.Start()```</small>
+- Removing Shell from inside Shiro
+    - <small>Now, instead of being part of Shiro, it can use Shiro's resources as an application.</small>
+- Removing unused ```bios``` class from ```cursor``` class
+
 ## [v0.3] - 2019-05-11
 
 - New namespaces, classes, methods, and functions